[refactor] 코드 패키지화 수정
