--- app/src/main/java/org/sopt/sample/data/remote/model/RequestLoginDTO.kt
@@ -1,12 +1,0 @@
-package org.sopt.sample.data.remote
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class RequestLoginDTO(
-    @SerialName("email")
-    val email: String,
-    @SerialName("password")
-    val password: String
-)
