+++ app/src/main/java/org/sopt/sample/presentation/login/view/LoginActivity.kt
@@ -5,14 +5,14 @@ import android.os.Bundle
 import android.widget.Toast
 import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
-import org.sopt.sample.data.remote.AuthService
-import org.sopt.sample.data.remote.RequestLoginDTO
-import org.sopt.sample.data.remote.ResponseLoginDTO
-import org.sopt.sample.data.remote.ServicePool
+import org.sopt.sample.data.remote.api.AuthService
+import org.sopt.sample.data.remote.api.ServicePool
+import org.sopt.sample.data.remote.model.RequestLoginDTO
+import org.sopt.sample.data.remote.model.ResponseLoginDTO
 import org.sopt.sample.databinding.ActivityLoginBinding
 import org.sopt.sample.presentation.login.viewmodel.LoginViewModel
 import org.sopt.sample.presentation.main.view.MainActivity
-import org.sopt.sample.presentation.signup.SignUpActivity
+import org.sopt.sample.presentation.signup.view.SignUpActivity
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
