+++ app/src/main/java/org/sopt/sample/presentation/main/adapter/FollowersAdapter.kt
@@ -1,11 +1,11 @@
-package org.sopt.sample.fragments.adapters
+package org.sopt.sample.presentation.main.adapter
 
 import android.content.Context
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
-import org.sopt.sample.data.remote.ResponseGetUsersDTO
+import org.sopt.sample.data.remote.model.ResponseGetUsersDTO
 import org.sopt.sample.databinding.ItemUsersBinding
 
 class FollowersAdapter(context: Context) : RecyclerView.Adapter<RecyclerView.ViewHolder>() {
