+++ app/src/main/java/org/sopt/sample/presentation/main/viewmodel/HomeViewModel.kt
@@ -1,7 +1,7 @@
 package org.sopt.sample.presentation.main.viewmodel
 
 import androidx.lifecycle.ViewModel
-import org.sopt.sample.data.remote.ResponseGetUsersDTO
+import org.sopt.sample.data.remote.model.ResponseGetUsersDTO
 
 class HomeViewModel : ViewModel() {
     var userList = listOf<ResponseGetUsersDTO.User>()
