+++ app/src/main/res/layout/activity_main.xml
@@ -8,7 +8,7 @@
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/fragmentContainerView"
-        android:name="org.sopt.sample.fragments.HomeFragment"
+        android:name="org.sopt.sample.presentation.main.fragment.HomeFragment"
         android:layout_width="match_parent"
         android:layout_height="0dp"
         app:layout_constraintBottom_toTopOf="@+id/bnv_main"
