+++ app/src/main/res/layout/fragment_search.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".fragments.SearchFragment">
+    tools:context=".presentation.main.fragment.SearchFragment">
 
     <TextView
         android:layout_width="wrap_content"
